app/requirements.txt
Flask
mysql-connector
flask_table